# openapi/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Energy Allocation API
  version: 1.0.0
  description: >-
    Tenant-aware API for onboarding facilities, ingesting data, generating model-driven
    recommendations, and tracking realized savings.
servers:
  - url: https://api.energyallocation.com
security:
  - oauth2: [client_admin, facility_manager, energy_analyst]
paths:
...
  /v1/facilities/{facility_id}/ingestion/imports:
    post:
      tags: [Ingestion]
      summary: Submit historical import
      operationId: submitImport
      parameters:
        - $ref: '#/components/parameters/facilityId'
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateImportRequest'
      responses:
        '202':
          description: Import accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportJob'
  /v1/ingestion/imports/{import_id}:
    get:
      tags: [Ingestion]
      summary: Get import status
      operationId: getImportStatus
      parameters:
        - $ref: '#/components/parameters/importId'
      responses:
        '200':
          description: Import progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportJob'
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://auth.energyallocation.com/oauth/token
          scopes:
            client_admin: Manage client-level resources
            facility_manager: Manage facility operations
            energy_analyst: Read analytics and model outputs
            viewer: Read-only access
            ops_admin: Internal operations access
            ml_engineer: Internal ML administration
  parameters:
...
    facilityId: { in: path, name: facility_id, required: true, schema: { type: string } }
...
    importId: { in: path, name: import_id, required: true, schema: { type: string } }
...
    idempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, minLength: 8 }
  schemas:
...
    CreateImportRequest:
      type: object
      properties:
        type: { type: string, enum: [csv, parquet, jsonl] }
        file_id: { type: string }
        schema_hint: { type: string }
        time_column: { type: string }
      required: [type, file_id, time_column]
    ImportJob:
      type: object
      properties:
        import_id: { type: string }
        status: { type: string, enum: [queued, running, completed, failed] }
        progress_pct: { type: number }
        estimated_rows: { type: integer }
        rows_processed: { type: integer }
        rows_rejected: { type: integer }
        started_at: { type: string, format: date-time }
        completed_at: { type: [string, 'null'], format: date-time }
      required: [import_id, status]
...